+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## [1.0.4](https://github.com/Rooknj/prysma-ui/compare/v1.0.3...v1.0.4) (2019-07-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency yargs to v13.3.0 ([5be029c](https://github.com/Rooknj/prysma-ui/commit/5be029c))
+
+## [1.0.3](https://github.com/Rooknj/prysma-ui/compare/v1.0.2...v1.0.3) (2019-07-15)
+
+
+### Bug Fixes
+
+* **deps:** pin dependency @material-ui/styles to 4.2.0 ([552b64e](https://github.com/Rooknj/prysma-ui/commit/552b64e))
+
+## [1.0.2](https://github.com/Rooknj/prysma-ui/compare/v1.0.1...v1.0.2) (2019-07-14)
+
+
+### Bug Fixes
+
+* **ci:** fixed deploy script ([0c9fa42](https://github.com/Rooknj/prysma-ui/commit/0c9fa42))
+* **ci:** release zipped assets ([2e1c050](https://github.com/Rooknj/prysma-ui/commit/2e1c050))
+
+## [1.0.1](https://github.com/Rooknj/prysma-ui/compare/v1.0.0...v1.0.1) (2019-07-14)
+
+
+### Bug Fixes
+
+* properly disabled color picker on disconnect ([48ce21c](https://github.com/Rooknj/prysma-ui/commit/48ce21c))
+
 # 1.0.0 (2019-07-14)